Add files via upload
This was part of the CrazyFlie drone team capstone project, the most important file and scripting to examine which explains a little bit about our reverse engineering/understanding of how the drone moves. The firmware is in C while the scripting is in Python and there's a lot of code translation and communication among the drone's components in order to achieve flight, obviously.